Moved CommandLineArgs to its own file.
